requirements.txt
Flask==2.3.3
Pillow==10.0.1
pandas==2.1.1
Werkzeug==2.3.7
openpyxl==3.1.2
xlrd==2.0.1
